refactor(repository, profile): profile repository 통합
