+++ src/main/java/com/pnu/momeet/domain/profile/repository/ProfileRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.Lock;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface ProfileRepository extends JpaRepository<Profile, UUID> {
+public interface ProfileRepository extends JpaRepository<Profile, UUID>, ProfileDslRepository {
     @EntityGraph(attributePaths = {"baseLocation"})
     Optional<Profile> findByMemberId(UUID memberId);
